Merge CRI Registries into config.toml
Because of https://github.com/containerd/containerd/issues/5837 we're
not able to seperate out the registries directive whilst using the
Nvidia runtime.

To workaround this we will edit the original config.toml so all
definitions are in-line. Even though this will likely result in
conflicts with upstream in the future.